+++ app/src/main/java/com/shall_we/home/ProductListFragment.kt
@@ -3,11 +3,14 @@ package com.shall_we.home
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
+import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
 import android.widget.AdapterView
 import android.widget.LinearLayout
 import android.widget.Spinner
+import androidx.activity.OnBackPressedCallback
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -30,24 +33,23 @@ class ProductListFragment : Fragment() , ProductAdapter.OnItemClickListener{
     override fun onItemClick(item: ProductData) {
         // 클릭된 아이템의 정보를 사용하여 다른 프래그먼트로 전환하는 로직을 작성
         val newFragment = ExperienceDetailFragment() // 전환할 다른 프래그먼트 객체 생성
-
-
         val bundle = Bundle()
         bundle.putInt("id", item.giftid) // 클릭된 아이템의 이름을 "title" 키로 전달
-
         newFragment.arguments = bundle
-
-
         // 프래그먼트 전환
         parentFragmentManager.beginTransaction()
-            .replace(R.id.nav_host_fragment, newFragment)
+            .add(R.id.nav_host_fragment, newFragment)
             .addToBackStack(null)
             .commit()
 
     }
 
 
-
+    override fun onResume() {
+        super.onResume()
+        val supportActionBar = (requireActivity() as AppCompatActivity).supportActionBar
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -133,6 +135,8 @@ class ProductListFragment : Fragment() , ProductAdapter.OnItemClickListener{
         })
 
         initSpinner(binding.spinner)
+
+
 
         return binding.root
     }
