+++ app/src/main/java/com/shall_we/home/RecomAdapter.kt
@@ -46,7 +46,7 @@ class RecomAdapter(private val context: Context) : RecyclerView.Adapter<RecomAda
                     fragment.arguments = bundle
                     val fragmentManager = (itemView.context as FragmentActivity).supportFragmentManager
                     fragmentManager.beginTransaction()
-                        .replace(R.id.nav_host_fragment, fragment)
+                        .add(R.id.nav_host_fragment, fragment)
                         .addToBackStack(null)
                         .commit()
                 }
