--- app/src/main/res/layout/fragment_home_realtime.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".home.HomeRealtimeFragment"
-    android:id="@+id/homeRealtime_layout">
-
-
-    <!-- TODO: Update blank fragment layout -->
-
-    <TextView
-        android:id="@+id/realtimeText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:text="실시간 인기 경험"
-        android:textColor="@color/black"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_category"
-        android:layout_width="0dp"
-        android:layout_height="28dp"
-        android:layout_marginTop="15dp"
-        android:layout_marginEnd="16dp"
-        android:orientation="horizontal"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@+id/realtimeText"
-        app:layout_constraintTop_toBottomOf="@+id/realtimeText"
-        tools:listitem="@layout/item_realtime_category" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_realtime"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="15dp"
-        android:layout_marginEnd="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@+id/rv_category"
-        app:layout_constraintTop_toBottomOf="@+id/rv_category"
-        tools:listitem="@layout/item_product" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
